r10460/sofa : Fix a bunch of "loss of data" warnings in framework/
Here starts the great journey of fixing the bazillion warnings due
to not using size_t. I took no risk and fixed only simple cases.


Former-commit-id: e74770d303293b688cee5f5392301103f5264b7b

+++ framework/sofa/core/behavior/Mass.inl
@@ -325,7 +325,7 @@ double Mass<DataTypes>::getElementMass(unsigned int ) const
 template <class DataTypes>
 void Mass<DataTypes>::getElementMass(unsigned int , defaulttype::BaseMatrix *m) const
 {
-    static unsigned int dimension = defaulttype::DataTypeInfo<Coord>::size();
+    static unsigned int dimension = (unsigned int) defaulttype::DataTypeInfo<Coord>::size();
     if (m->rowSize() != dimension || m->colSize() != dimension) m->resize(dimension,dimension);
 
     m->clear();
